# backend-gin/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/hello-world/user": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hello World"
                ],
                "summary": "Hello World Endpoint for authenticated users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.HelloUserResponse"
                        }
                    }
                }
            }
        },
...
        "/api/users": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User creation",
                "parameters": [
                    {
                        "type": "string",
                        "example": "mslimbeji@gmail.com",
                        "description": "The user email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "example": false,
                        "description": "Whether the user is an admin or not",
                        "name": "isAdmin",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minLength": 2,
                        "type": "string",
                        "example": "Slim Beji",
                        "description": "The user name, two characters at least",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minLength": 8,
                        "type": "string",
                        "example": "very_secret",
                        "description": "The user password, 8 characters at least",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "the user profile image",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UserRead"
                        }
                    }
                }
            }
        },
        "/api/users/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Search and Filter users",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "eq:mslimbeji@gmail.com"
                        ],
                        "description": "The user email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "id",
                                "name",
                                "email",
                                "isAdmin",
                                "imageUrl",
                                "places",
                                "createdAt"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "id",
                            "place"
                        ],
                        "description": "Fields to include in the response; omit for full document",
                        "name": "fields",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "123456789"
                        ],
                        "description": "The user ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "eq:Slim Beji"
                        ],
                        "description": "The user name, two characters at least",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "The page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 100,
                        "description": "Items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "createdAt",
                                "-createdAt",
                                "name",
                                "-name",
                                "email",
                                "-email"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "-createdAt"
                        ],
                        "description": "Fields to use for sorting. Use the '-' for descending sorting",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UsersPaginated"
                        }
                    }
                }
            }
        },
        "/api/users/search": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Search and Retrieve users",
                "parameters": [
                    {
                        "description": "POST parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UserSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UsersPaginated"
                        }
                    }
                }
            }
        },
        "/api/users/{userId}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Search and Retrieve user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UserRead"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PUT parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UserPut"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UserRead"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.UserDeleteResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "routes.HelloUserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello Slim"
                }
            }
        },
...
        "routes.UserDeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted user 507f1f77bcf86cd799439011"
                }
            }
        },
...
        "schemas.Location": {
            "type": "object",
            "properties": {
                "lat": {
                    "description": "The latitude of the place",
                    "type": "number",
                    "example": 51.48180425016331
                },
                "lng": {
                    "description": "The longitude of the place",
                    "type": "number",
                    "example": -0.19090418688755467
                }
            }
        },
...
        "schemas.PlaceRead": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "The place address",
                    "type": "string",
                    "minLength": 1,
                    "example": "Fulham road"
                },
                "createdAt": {
                    "description": "creation datetime                                            // last update datetime",
                    "type": "string",
                    "example": "2024-01-12T10:15:30.000Z"
                },
                "creatorId": {
                    "description": "The ID of the place creator",
                    "type": "integer",
                    "example": 123456789
                },
                "description": {
                    "description": "The place description, 10 characters minimum",
                    "type": "string",
                    "minLength": 10,
                    "example": "Stadium of Chelsea football club"
                },
                "id": {
                    "description": "The ID of the place",
                    "type": "integer",
                    "example": 123456789
                },
                "imageUrl": {
                    "description": "local url on the storage",
                    "type": "string",
                    "example": "avatar2_80e32f88-c9a5-4fcd-8a56-76b5889440cd.jpg"
                },
                "location": {
                    "description": "Location object (can be sent as JSON string)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.Location"
                        }
                    ]
                },
                "title": {
                    "description": "The place title/name, 10 characters minimum",
                    "type": "string",
                    "minLength": 10,
                    "example": "Stamford Bridge"
                }
            }
        },
...
        "schemas.UserPlace": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "The place address",
                    "type": "string",
                    "minLength": 1,
                    "example": "Fulham road"
                },
                "id": {
                    "description": "The ID of the place",
                    "type": "integer",
                    "example": 123456789
                },
                "title": {
                    "description": "The place title/name, 10 characters minimum",
                    "type": "string",
                    "minLength": 10,
                    "example": "Stamford Bridge"
                }
            }
        },
        "schemas.UserPut": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "The user email",
                    "type": "string",
                    "example": "mslimbeji@gmail.com"
                },
                "name": {
                    "description": "The user name, two characters at least",
                    "type": "string",
                    "minLength": 2,
                    "example": "Slim Beji"
                },
                "password": {
                    "description": "The user password, 8 characters at least",
                    "type": "string",
                    "minLength": 8,
                    "example": "very_secret"
                }
            }
        },
        "schemas.UserRead": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "creation datetime                                            // last update datetime",
                    "type": "string",
                    "example": "2024-01-12T10:15:30.000Z"
                },
                "email": {
                    "description": "The user email",
                    "type": "string",
                    "example": "mslimbeji@gmail.com"
                },
                "id": {
                    "description": "The user ID",
                    "type": "integer",
                    "example": 12345678
                },
                "imageUrl": {
                    "description": "local url on the storage",
                    "type": "string",
                    "example": "avatar2_80e32f88-c9a5-4fcd-8a56-76b5889440cd.jpg"
                },
                "isAdmin": {
                    "description": "Whether the user is an admin or not",
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "description": "The user name, two characters at least",
                    "type": "string",
                    "minLength": 2,
                    "example": "Slim Beji"
                },
                "places": {
                    "description": "Places created by the user",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.UserPlace"
                    }
                }
            }
        },
        "schemas.UserSearch": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "The user email",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "eq:mslimbeji@gmail.com"
                    ]
                },
                "fields": {
                    "description": "Fields to include in the response; omit for full document",
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "id",
                            "name",
                            "email",
                            "isAdmin",
                            "imageUrl",
                            "places",
                            "createdAt"
                        ]
                    },
                    "example": [
                        "id",
                        "place"
                    ]
                },
                "id": {
                    "description": "The user ID",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "123456789"
                    ]
                },
                "name": {
                    "description": "The user name, two characters at least",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "eq:Slim Beji"
                    ]
                },
                "page": {
                    "description": "The page number",
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                },
                "size": {
                    "description": "Items per page",
                    "type": "integer",
                    "default": 100,
                    "maximum": 100,
                    "minimum": 1
                },
                "sort": {
                    "description": "Fields to use for sorting. Use the '-' for descending sorting",
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "createdAt",
                            "-createdAt",
                            "name",
                            "-name",
                            "email",
                            "-email"
                        ]
                    },
                    "example": [
                        "-createdAt"
                    ]
                }
            }
        },
        "schemas.UsersPaginated": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.PlaceRead"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "totalCount": {
                    "type": "integer",
                    "example": 40
                },
                "totalPages": {
                    "type": "integer",
                    "example": 2
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/auth/signin"
        }
    },
...
}